backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
python-decouple==3.8
Pillow==10.1.0
django-extensions==3.2.3
sentry-sdk==1.38.0
gunicorn==21.2.0